+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added support for *prev-ver* flag in **update-release-notes** command.
 * Improved retry support when building docker images for linting.
 * Added the *--skip-id-set-creation* flag to **validate** command in order to add the capability to run validate command without creating id_set validation.
+* Added version of availability (fromversion) to the README and to the release notes, when a new integration is released, in **update-release-notes** and **generate_docs** command.
 
 # 1.2.11
 * Fixed an issue where the ***generate-docs*** command reset the enumeration of line numbering after an MD table.